Fixing cmake build scripts to support 64bit on Windows. Also refactoring some pointer handling to support 64bit on Windows (Linux and Mac have always supported it). Adding instructions for setting up a Windows 64 bit build environment using MSYS2.

+++ include/Qt/VideoRenderer.h
@@ -45,7 +45,7 @@ public:
     ~VideoRenderer();
 
     /// Override QWidget which needs to be painted
-    void OverrideWidget(long qwidget_address);
+    void OverrideWidget(long long qwidget_address);
 
 signals:
 	void present(const QImage &image);
